Reorganize context files to use index - Update imports
